inbox_util: Move is_in_focus to inbox_ui.
